My final project, submitted to the instructor 12/9/16
The final project in its entirety. The instructor required us to use rmarkdown to publish our code, which is why this is an HTML file. This file has interactive links that can be used when opened in browser, and contains the RMD source code file which could be ran and reproduced on any R console (given that said console includes specific installations and options).